add task 0

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,9 +4,9 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
-	"""generates a random value between 0 and max_delay and uses it as a delay
-	after delay returns the value"""
-	delay = random.uniform(0, max_delay)
-	await asyncio.sleep(delay)
-	return delay
+async def wait_random(max_delay: int = 10) -> float:
+    """generates a random value between 0 and max_delay and uses it as a delay
+    after delay returns the value"""
+    delay = random.uniform(0, max_delay)
+    await asyncio.sleep(delay)
+    return delay
